+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/service/AsyncDataApiService.kt
@@ -99,7 +99,7 @@ class AsyncDataApiService(
         reportOrDashboardId = productDefinition.report.id,
         reportOrDashboardName = productDefinition.report.name,
         preGeneratedDatasetTableId = preGeneratedTableId,
-        multiphaseQuery = productDefinition.reportDataset.multiphaseQuery,
+        multiphaseQueries = productDefinition.reportDataset.multiphaseQuery,
       )
   }
 
@@ -137,7 +137,7 @@ class AsyncDataApiService(
         productDefinitionName = productDefinition.name,
         reportOrDashboardId = productDefinition.dashboard.id,
         reportOrDashboardName = productDefinition.dashboard.name,
-        multiphaseQuery = productDefinition.dashboardDataset.multiphaseQuery,
+        multiphaseQueries = productDefinition.dashboardDataset.multiphaseQuery,
       )
   }